requirements.txt
accelerate==1.1.1
aenum==3.1.15
affine==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
bitsandbytes==0.45.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
diffusers==0.32.1
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fiona==1.10.1
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.4
jsonargparse==4.36.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightly==1.5.17
lightly-utils==0.0.2
lightning==2.5.0.post0
lightning-utilities==0.11.8
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
regex==2024.11.6
requests==2.32.3
rich==13.9.4
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.15.0
segmentation_models_pytorch==0.4.0
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboardX==2.6.2.2
tifffile==2024.12.12
timm==1.0.14
tokenizers==0.20.3
torch==2.5.1
torch-fidelity==0.3.0
torchgeo==0.6.2
torchmetrics==1.6.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
xformers==0.0.29.post1
yarl==1.17.2
zipp==3.21.0
